requirements.txt
Pillow==8.1.2
RPi.GPIO==0.7.1
spidev==3.6